requirement.txt
django-extensions==2.1.3
django-phonenumber-field==2.0.1
phonenumbers==8.9.14
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
django-autoslug==1.9.3